requirements.txt
wcmatch==5.0
requests
pyyaml
